Delete unnecessary chat components and put all together in Chat.js
