+++ frontend/src/App.js
@@ -4,8 +4,8 @@ import Register from "./components/Register";
 import Login from "./components/Login";
 import Home from "./components/Home";
 import Card from "./components/Card";
-import ChatWrapper from "./components/chatting/ChatWrapper";
 import MainNavbar from "./components/MainNavbar";
+import Chat from "./components/Chat";
 
 function App() {
   return (
@@ -15,7 +15,7 @@ function App() {
       <Route path="/login" element={<Login />} />
       <Route path="/card" element={<Card />} />
       <Route path="/main" element={<MainNavbar />} />
-      <Route path="/main/chat" element={<ChatWrapper />} />
+      <Route path="/chat" element={<Chat />} />
     </Routes>
   );
 }
