--- frontend/src/components/chatting/Sidebar.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Navbar from "./Navbar";
-import Chats from "./Chats";
-
-function Sidebar() {
-  return (
-    <div className="sidebar">
-      <Navbar />
-      <Chats />
-    </div>
-  );
-}
-
-export default Sidebar;
