--- frontend/src/components/chatting/SingleMessage.js
@@ -1,16 +1,0 @@
-import React from "react";
-
-function SingleMessage() {
-  return (
-    <div className="message owner">
-      <div className="messageInfo">
-        <img src="../img/chat_01.jpg" alt="chatpic" />
-      </div>
-      <div className="messageContent">
-        <p>hello</p>
-      </div>
-    </div>
-  );
-}
-
-export default SingleMessage;
